+++ src/components/user-component.js
@@ -4,8 +4,8 @@ import {Link} from "react-router-dom";
 
 
 const username = "User name";
-const user_avatar = "";
-const user_backgroun_image = "";
+// const user_avatar = "";
+// const user_backgroun_image = "";
 const user_repository_count = "User repository count ";
 
 
@@ -19,15 +19,20 @@ class UserComponent extends Component {
                     <div className="card card-inverse">
                         <div className="card-img-overlay">
                             <div className="flex justify-content-between">
-                                <Link to='/user_settings'><button type="button" className="btn btn-light btn-sm">Settings</button></Link>
-                                <Link to='/'><button type="button" className="btn btn-light btn-sm">Sign out</button></Link>
+                                <Link to='/user_settings'>
+                                    <button type="button" className="btn btn-light btn-sm">Settings</button>
+                                </Link>
+                                <Link to='/'>
+                                    <button type="button" className="btn btn-light btn-sm">Sign out</button>
+                                </Link>
                             </div>
                         </div>
                     </div>
                     <CardImg top width="100%"  /*src={user_background_image}*/ src="https://picsum.photos/300/180"
                              alt="Card image cap"/>
                     <CardBody>
-                        <Link to='/user'><img /*src = {user_avatar}*/ src="https://picsum.photos/150/150" alt="Avatar" className="br-50"/></Link>
+                        <Link to='/user'><img /*src = {user_avatar}*/ src="https://picsum.photos/150/150" alt="Avatar"
+                                                                      className="br-50"/></Link>
                         <CardTitle>{username}</CardTitle>
                         <CardText>{user_repository_count}</CardText>
                     </CardBody>
