+++ src/components/user-page.js
@@ -1,4 +1,4 @@
-import React, {Component} from 'react';
+import React from 'react';
 import RepoTabs from './repository-list'
 import Page from './page';
