+++ src/components/user-settings-page.js
@@ -1,4 +1,4 @@
-import React, {Component} from 'react';
+import React from 'react';
 import UserSettingsForm from './user-settings-form';
 import Page from './page';
