+++ src/constants/about-devs.js
@@ -1,0 +1,46 @@
+const devs = [
+    {
+        name: 'Name Surname',
+        description: 'Lorem ipsum dolor sit amet, consectetur adipisicing elit. Esse, sit!',
+        avatar: 'https://picsum.photos/150/150',
+        bgimg: 'https://picsum.photos/300/180',
+    },
+    {
+        name: 'Name Surname',
+        description: 'Lorem ipsum dolor sit amet, consectetur adipisicing elit. Esse, sit!',
+        avatar: 'https://picsum.photos/150/150',
+        bgimg: 'https://picsum.photos/300/180',
+    },
+    {
+        name: 'Name Surname',
+        description: 'Lorem ipsum dolor sit amet, consectetur adipisicing elit. Esse, sit!',
+        avatar: 'https://picsum.photos/150/150',
+        bgimg: 'https://picsum.photos/300/180',
+    },
+    {
+        name: 'Name Surname',
+        description: 'Lorem ipsum dolor sit amet, consectetur adipisicing elit. Esse, sit!',
+        avatar: 'https://picsum.photos/150/150',
+        bgimg: 'https://picsum.photos/300/180',
+    },
+    {
+        name: 'Name Surname',
+        description: 'Lorem ipsum dolor sit amet, consectetur adipisicing elit. Esse, sit!',
+        avatar: 'https://picsum.photos/150/150',
+        bgimg: 'https://picsum.photos/300/180',
+    },
+    {
+        name: 'Name Surname',
+        description: 'Lorem ipsum dolor sit amet, consectetur adipisicing elit. Esse, sit!',
+        avatar: 'https://picsum.photos/150/150',
+        bgimg: 'https://picsum.photos/300/180',
+    },
+    {
+        name: 'Name Surname',
+        description: 'Lorem ipsum dolor sit amet, consectetur adipisicing elit. Esse, sit!',
+        avatar: 'https://picsum.photos/150/150',
+        bgimg: 'https://picsum.photos/300/180',
+    },
+];
+
+export default devs;
